+++ test/sample_prompts/prompt_hello_world.py
@@ -16,4 +16,3 @@ class PromptHelloWorldSystem(PromptBase):
 
     def set_prompt(self) -> str:
         return "Hello {name}"
-